requirements.txt
requests
beautifulsoup4
lxml
readability-lxml
python-dotenv
